// src/app/shared/utils/enums/atoms-values.ts
export enum ButtonSizes {
  LARGE = 'large',
  MEDIUM = 'medium',
  SMALL = 'small',
}

export enum ButtonTypes {
  MAIN = 'main',
  SECONDARY = 'secondary',
  SUBMIT = 'submit',
  DANGER = 'danger',
}

export enum InputSizes {
  NORMAL = 'normal',
  SMALL = 'small',
  LARGE = 'large',
}

export enum LogoSizes {
  LARGE = 'large',
  MEDIUM = 'medium',
  SMALL = 'small',
}